Plus aucun avertissement clippy
Un seul corrige un vrai defaut visible : `{:#02X}` n'alignait rien, le
prefixe "0x" ajoute par `#` comptant dans la largeur demandee. La commande
"s" du moniteur affichait donc 0x5 la ou le reste ecrit 0x05. Passe en
`{:#04X}`, verifie a l'execution.

Le reste est de la forme :
- quatre `if` imbriques (fdc.rs x2, machine.rs, memory.rs) fusionnes en
  chaines `let`, disponibles en edition 2024 ;
- le triplet (MonitorCmd, String, String) qui circule sur le canal devient
  le type nomme monitor::MonitorMessage, partage entre console.rs et
  machine.rs — c'est le point d'entree unique decrit dans Plan V2.md, il
  meritait un nom ;
- io::Error::other(e) au lieu de io::Error::new(ErrorKind::Other, e) ;
- `let _ =` sur un canvas.clear() qui ne renvoie rien, push_str d'un
  caractere seul, et un `as i32` sur une valeur deja i32.

118 tests, zero avertissement clippy (12 avant).

+++ src/video.rs
@@ -260,7 +260,7 @@ fn render_byte_mode2(
 ) {
     for i in 0..8 {
         let color_idx = (byte >> (7 - i)) & 1;
-        draw_pixel_run(state, frame_buffer, color_idx, x_byte + i as i32, 1, y);
+        draw_pixel_run(state, frame_buffer, color_idx, x_byte + i, 1, y);
     }
 }
